Se implementan exceptions para cuando se intenta actualizar o eliminar una transacción que no existe en la base de datos

+++ src/main/java/com/drewcode/moneyflow/service/TransactionService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.drewcode.moneyflow.entity.Transaction;
+import com.drewcode.moneyflow.exception.TransactionNotFoundException;
 import com.drewcode.moneyflow.repository.TransactionRepository;
 
 @Service
@@ -30,10 +31,15 @@ public class TransactionService {
 
     public Transaction findById(Long id) {
         return repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Transaction not found"));
+                .orElseThrow(() -> new TransactionNotFoundException(id));
     }
 
     public void deleteById(Long id) {
+        boolean exists = repository.existsById(id);
+
+        if (!exists) {
+            throw new TransactionNotFoundException(id); // tu excepción personalizada
+        }
         repository.deleteById(id);
     }
